Updated search bar config to hide/show it
