+++ android/geowidget/src/main/java/org/smartregister/fhircore/geowidget/screens/GeoWidgetFragment.kt
@@ -93,28 +93,13 @@ class GeoWidgetFragment : Fragment() {
   }
 
   private fun setupViews(): LinearLayout {
-    val toolbar = setUpToolbar()
+    //val toolbar = setUpToolbar()
     mapView = setUpMapView()
 
     return LinearLayout(requireContext()).apply {
       orientation = LinearLayout.VERTICAL
-      addView(toolbar)
+      //addView(toolbar)
       addView(mapView)
-    }
-  }
-
-  // TODO: move toolbar to quest
-  private fun setUpToolbar(): Toolbar {
-    return Toolbar(requireContext()).apply {
-      popupTheme = R.style.AppTheme
-      visibility = View.VISIBLE
-      navigationIcon =
-        ContextCompat.getDrawable(context, androidx.appcompat.R.drawable.abc_ic_ab_back_material)
-      layoutParams = LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, 168)
-      setBackgroundColor(ContextCompat.getColor(requireContext(), R.color.colorPrimary))
-      setNavigationOnClickListener {
-        activity?.onBackPressedDispatcher?.onBackPressed()
-      }
     }
   }
